+++ src/modules/auth/dto/update-auth.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateAuthDto } from './create-auth.dto';
+import { CreateAuthDto } from './register.dto';
 
 export class UpdateAuthDto extends PartialType(CreateAuthDto) {}
